panel/backend/requirements.txt
# Panel backend - MTProxy management
fastapi>=0.115.0,<0.116
uvicorn[standard]>=0.32.0,<0.33
sqlalchemy>=2.0.0,<3
alembic>=1.14.0,<2
pydantic>=2.0,<3
pydantic-settings>=2.0,<3
PyJWT>=2.8.0,<3
bcrypt>=4.0.0,<5
python-multipart>=0.0.12
httpx>=0.28.0,<0.29
apscheduler>=3.10.0,<4
tomli-w>=1.0.0
# tests
pytest>=8.0
pytest-asyncio>=0.24
httpx>=0.28.0
